--- Assets/Kojima.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6a0e3f794c9ad534fa2f56664599598e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
