--- Assets/Kojima/Kojima.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 749a6fb0c258c644f8b474c038dd448a
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
