--- Assets/Kojima/Matelial.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ce8746c4bcdce864ebe0d840ee208892
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
