--- Assets/Kojima/Matelial/Kship.physicsMaterial2D.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 44c4242eb08d70d4fa310add87f41df7
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
