--- Assets/Kojima/Resources.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9d55923087cdebf4dbebb73f1b893856
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
