--- Assets/Kojima/Resources/KojiShip.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: caf354567c4f2e240890aba87ef41572
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
