--- Assets/Kojima/Tanaka 1.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: f22bf9739bd1b7946a79ce22041664b6
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
