--- Assets/Kojima/sclipts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 822d6ae45dcf53f458d1f84f00a7c120
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
