--- Assets/Kojima/sclipts/KojiShipController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f6446d3dd97c13d4bbba31bb058a8135
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
