+++ Assets/Resources/SpaceShip.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7d8d315d2cf70614695fc85a2e53abd3
+guid: caf354567c4f2e240890aba87ef41572
 PrefabImporter:
   externalObjects: {}
   userData:
