--- Assets/Scripts/tanaka/GitTest.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b520a075f30d523428da9e28144932eb
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
